+++ FromLocalsToLocals/Controllers/HomeController.cs
@@ -125,61 +125,5 @@ namespace FromLocalsToLocals.Controllers
 
             return LocalRedirect(returnUrl);
         }
-    
-        [HttpGet]
-        public async Task<IActionResult> NewsFeed(PostVM model)
-        {
-
-            if (model.User == null)
-            {
-                model.User = await _userManager.Users.FirstOrDefaultAsync(x => x.Id == _userManager.GetUserId(User));                
-            }
-            if (model.ActiveTab == null)
-            {
-                model.ActiveTab = Tab.AllFeed;
-            }
-
-            return View(model);
-        }
-
-        [Authorize]
-        public IActionResult SwitchTabs(PostVM model,string tabName)
-        {
-            model.ActiveTab = tabName.ParseEnum<Tab>();
-            return RedirectToAction(nameof(HomeController.NewsFeed), model);
-        }
-
-        [HttpPost]
-        [Authorize]
-        public async Task<IActionResult> CreatePost(PostVM model)
-        {
-            if (!ModelState.IsValid)
-            {
-                _toastNotification.AddErrorToastMessage("Cannot create post with empty message");
-                return Redirect(model.PostBackUrl);
-            }
-
-            var userId = _userManager.GetUserId(User);
-            model.SelectedVendor = await _context.Vendors.FirstOrDefaultAsync(x => x.Title == model.SelectedVendorTitle && x.UserID == userId);
-
-            if (model.SelectedVendor != null)
-            {
-                try
-                {
-                    _context.Posts.Add(new Post(model));
-                    await _context.SaveChangesAsync();
-                }
-                catch (Exception)
-                {
-                    _toastNotification.AddErrorToastMessage("Something unexpected happened. Cannot create a post.");
-                    return Redirect(model.PostBackUrl);
-                }
-            }
-
-            model.PostText = "";
-
-            return Redirect(model.PostBackUrl);
-        }
-
     }
 }
