+++ FromLocalsToLocals/Database/AppDbContext.cs
@@ -27,7 +27,7 @@ namespace FromLocalsToLocals.Database
                 .HasKey(bc => new { bc.UserID, bc.VendorID });
             modelBuilder.Entity<Follower>()
               .HasOne(u => u.User)
-              .WithMany(u => u.Folllowing)
+              .WithMany(u => u.Following)
               .HasForeignKey(bc => bc.UserID).IsRequired().OnDelete(DeleteBehavior.Restrict);
             modelBuilder.Entity<Follower>()
                 .HasOne(bc => bc.Vendor)
