+++ FromLocalsToLocals/Views/Shared/Components/NewsFeed/Default.cshtml
@@ -4,169 +4,9 @@
 
 
 <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css" rel="stylesheet">
+<link rel="stylesheet" href="~/css/Posts.css">
+
 <div class="container">
-    <style>
-
-
-
-        .timeline {
-            list-style-type: none;
-            margin: 0;
-            padding: 0;
-            position: relative
-        }
-
-            .timeline:before {
-                content: '';
-                position: absolute;
-                top: 5px;
-                bottom: 5px;
-                width: 5px;
-                background: #2d353c;
-                left: 20%;
-                margin-left: -2.5px
-            }
-
-            .timeline > li {
-                position: relative;
-                min-height: 50px;
-                padding: 20px 0
-            }
-
-            .timeline .timeline-time {
-                position: absolute;
-                left: 0;
-                width: 18%;
-                text-align: right;
-                top: 30px
-            }
-
-                .timeline .timeline-time .date,
-                .timeline .timeline-time .time {
-                    display: block;
-                    font-weight: 600
-                }
-
-                .timeline .timeline-time .date {
-                    line-height: 16px;
-                    font-size: 12px
-                }
-
-                .timeline .timeline-time .time {
-                    line-height: 24px;
-                    font-size: 20px;
-                    color: #242a30
-                }
-
-            .timeline .timeline-icon {
-                left: 15%;
-                position: absolute;
-                width: 10%;
-                text-align: center;
-                top: 40px
-            }
-
-                .timeline .timeline-icon a {
-                    text-decoration: none;
-                    width: 20px;
-                    height: 20px;
-                    display: inline-block;
-                    border-radius: 20px;
-                    background: #d9e0e7;
-                    line-height: 10px;
-                    color: #fff;
-                    font-size: 14px;
-                    border: 5px solid #2d353c;
-                    transition: border-color .2s linear
-                }
-
-            .timeline .timeline-body {
-                margin-left: 23%;
-                margin-right: 17%;
-                background: #fff;
-                position: relative;
-                padding: 20px 25px;
-                border-radius: 6px
-            }
-
-                .timeline .timeline-body:before {
-                    content: '';
-                    display: block;
-                    position: absolute;
-                    border: 10px solid transparent;
-                    border-right-color: #fff;
-                    left: -20px;
-                    top: 20px
-                }
-
-                .timeline .timeline-body > div + div {
-                    margin-top: 15px
-                }
-
-                    .timeline .timeline-body > div + div:last-child {
-                        margin-bottom: -20px;
-                        padding-bottom: 20px;
-                        border-radius: 0 0 6px 6px
-                    }
-
-        .timeline-header {
-            padding-bottom: 10px;
-            border-bottom: 1px solid #e2e7eb;
-            line-height: 30px
-        }
-
-            .timeline-header .userimage {
-                float: left;
-                width: 34px;
-                height: 34px;
-                border-radius: 40px;
-                overflow: hidden;
-                margin: -2px 10px -2px 0
-            }
-
-            .timeline-header .username {
-                font-size: 16px;
-                font-weight: 600
-            }
-
-                .timeline-header .username,
-                .timeline-header .username a {
-                    color: #2d353c
-                }
-
-        .timeline img {
-            max-width: 100%;
-            display: block
-        }
-
-        .timeline-content {
-            letter-spacing: .25px;
-            line-height: 18px;
-            font-size: 13px
-        }
-
-            .timeline-content:after,
-            .timeline-content:before {
-                content: '';
-                display: table;
-                clear: both
-            }
-
-        .timeline-footer {
-            background: #fff;
-            border-top: 1px solid #e2e7ec;
-            padding-top: 15px
-        }
-
-            .timeline-footer a:not(.btn) {
-                color: #575d63
-            }
-
-                .timeline-footer a:not(.btn):focus,
-                .timeline-footer a:not(.btn):hover {
-                    color: #2d353c
-                }
-    </style>
     <div class="row">
         <div class="col-md-12">
             @if (Model.User != null && Model.User.Vendors.Count > 0)
@@ -182,7 +22,7 @@
                                 <!-- begin timeline -->
                                 <ul class="timeline">
                                     <li>
-                                        <form asp-action="CreatePost" asp-controller="Home">
+                                        <form asp-action="CreatePost" asp-controller="Feed">
                                             <!-- end timeline-icon -->
                                             <!-- begin timeline-body -->
                                             <div class="timeline-body">
