+++ FromLocalsToLocals/Views/Shared/_Layout.cshtml
@@ -67,7 +67,7 @@
                         <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="FAQ">@_localizer["FAQ"]</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="NewsFeed">News Feed</a>
+                        <a class="nav-link text-dark" asp-area="" asp-controller="Feed" asp-action="NewsFeed">News Feed</a>
                     </li>
                 </ul>
                 @if (signInManager.IsSignedIn(User))
