+++ FromLocalsToLocals/Views/Vendors/Details.cshtml
@@ -82,7 +82,7 @@
             </div>
 
             <div class="row mt-3 mb-3" id="feed" style="display:none">
-                @await Component.InvokeAsync("NewsFeed", new FromLocalsToLocals.Models.ViewModels.PostVM { ActiveTab = FromLocalsToLocals.Models.ViewModels.Tab.VendorFeed, DisplayInDetails = true, SelectedVendor = Model, User = userForFeed});
+                @await Component.InvokeAsync("NewsFeed", new FromLocalsToLocals.Models.ViewModels.PostVM { ActiveTab = FromLocalsToLocals.Models.ViewModels.Tab.VendorFeed, DisplayInDetails = true, SelectedVendor = Model, User = userForFeed})
             </div>
 
         </div>
